Move implementations to core package
