+++ logbook-core/src/main/java/org/zalando/logbook/core/attributes/CompositeAttributeExtractor.java
@@ -1,10 +1,12 @@
-package org.zalando.logbook.attributes;
+package org.zalando.logbook.core.attributes;
 
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.extern.slf4j.Slf4j;
 import org.zalando.logbook.HttpRequest;
 import org.zalando.logbook.HttpResponse;
+import org.zalando.logbook.attributes.AttributeExtractor;
+import org.zalando.logbook.attributes.HttpAttributes;
 
 import javax.annotation.Nonnull;
 import java.util.HashMap;
