+++ logbook-core/src/main/java/org/zalando/logbook/core/attributes/JwtBaseExtractor.java
@@ -1,4 +1,4 @@
-package org.zalando.logbook.attributes;
+package org.zalando.logbook.core.attributes;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -8,6 +8,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.apiguardian.api.API;
 import org.zalando.logbook.HttpHeaders;
 import org.zalando.logbook.HttpRequest;
+import org.zalando.logbook.attributes.AttributeExtractor;
 
 import javax.annotation.Nonnull;
 import java.util.Base64;
